nonblockin streamin wip

+++ lib/krad_io/krad_file.c
@@ -234,7 +234,7 @@ kr_file_t *kr_file_open (char *path) {
   return ofile;
 }
 
-int kr_file_read (kr_file_t *file, void *buffer, size_t len) {
+ssize_t kr_file_read (kr_file_t *file, void *buffer, size_t len) {
   return read (file->fd, buffer, len);
 }
